@@ -1,10 +1,9 @@
 using System.Text.Json;
 
-namespace Api.DtoModels
+namespace Api.DtoModels;
+
+public record DtoPageData
 {
-    public record DtoPageData
-    {
-        public string Key { get; init; }
-        public JsonDocument Data { get; init; }
-    }
+    public string Key { get; init; }
+    public JsonDocument Data { get; init; }
 }
