@@ -1,10 +1,9 @@
 ﻿using System.Text.Json;
 
-namespace Api.Models.Entities
+namespace Api.Models.Entities;
+
+public class PageData
 {
-    public class PageData
-    {
-        public string Key { get; set; }
-        public JsonDocument Data { get; set; }
-    }
+    public string Key { get; set; }
+    public JsonDocument Data { get; set; }
 }
