@@ -1,10 +1,6 @@
-using System.Threading.Tasks;
-using Microsoft.AspNetCore.Http;
-using Microsoft.Extensions.Logging;
-namespace Api.Services
+namespace Api.Services;
+
+public interface IAssetFileService
 {
-    public interface IAssetFileService
-    {
-        Task<string> StoreAsset(IFormFile image);
-    }
+    Task<string> StoreAsset(IFormFile image);
 }
