@@ -1,47 +1,37 @@
-using System.Security.Claims;
-using Api.Controllers;
-using Api.DataAccess;
-using FluentAssertions;
-using Microsoft.AspNetCore.Http;
-using Microsoft.AspNetCore.Mvc;
-using Microsoft.Extensions.Logging.Abstractions;
-using NUnit.Framework;
+namespace Tests.ControllerTests;
 
-namespace Tests.Controller
+public class AdminControllerTest
 {
-    public class AdminControllerTest
-    {
-        private AdminController adminController;
-        private ClaimsPrincipal adminUser;
+    private AdminController adminController;
+    private ClaimsPrincipal adminUser;
 
-        public static AdminController GetAdminController()
-        {
-            var context = TestHelpers.CreateContext();
-            var donationRepository = new DonationRepository(context, TestHelpers.AspenMapper);
-            var eventRepository = new EventRepository(context, TestHelpers.AspenMapper);
-            return new AdminController(donationRepository, eventRepository, TestHelpers.AspenMapper);
-        }
-        
-        [SetUp]
-        public void Setup()
-        {
-            adminController = GetAdminController();
-            var userClaims = new Claim[] {
+    public static AdminController GetAdminController()
+    {
+        var context = TestHelpers.CreateContext();
+        var donationRepository = new DonationRepository(context, TestHelpers.AspenMapper);
+        var eventRepository = new EventRepository(context, TestHelpers.AspenMapper);
+        return new AdminController(donationRepository, eventRepository, TestHelpers.AspenMapper);
+    }
+
+    [SetUp]
+    public void Setup()
+    {
+        adminController = GetAdminController();
+        var userClaims = new Claim[] {
                 new Claim(ClaimTypes.NameIdentifier, "testAdmin"),
                 new Claim(ClaimTypes.Role, "admin-aspen")
             };
-            adminUser = new ClaimsPrincipal(
-                new ClaimsIdentity(userClaims, "TestAuthentication")
-            );
-        }
+        adminUser = new ClaimsPrincipal(
+            new ClaimsIdentity(userClaims, "TestAuthentication")
+        );
+    }
 
-        [Test]
-        public void CanGetFromAdminController()
-        {
-            adminController.ControllerContext = new ControllerContext();
-            adminController.ControllerContext.HttpContext = new DefaultHttpContext { User = adminUser };
-            var userClaims = adminController.Get();
-            userClaims.Should().NotBeEmpty();
-        }
+    [Test]
+    public void CanGetFromAdminController()
+    {
+        adminController.ControllerContext = new ControllerContext();
+        adminController.ControllerContext.HttpContext = new DefaultHttpContext { User = adminUser };
+        var userClaims = adminController.Get();
+        userClaims.Should().NotBeEmpty();
     }
 }
