@@ -1,34 +1,25 @@
-using System.Threading.Tasks;
-using Api.Services;
-using Microsoft.AspNetCore.Authorization;
-using Microsoft.AspNetCore.Http;
-using Microsoft.AspNetCore.Mvc;
-using Microsoft.Extensions.Logging;
-using Swashbuckle.AspNetCore.Annotations;
+namespace Api.Controllers;
 
-namespace Api.Controller
+public record Response<T> { public T Data { get; init; } }
+
+[ApiController]
+[Authorize]
+[Route("/api/[controller]")]
+public class AssetController : ControllerBase
 {
-    public record Response<T> { public T Data { get; init; } }
+    public IAssetFileService assetsFileService { get; }
 
-    [ApiController]
-    [Authorize]
-    [Route("/api/[controller]")]
-    public class AssetController : ControllerBase
+    public AssetController(IAssetFileService assetsFileService)
     {
-        public IAssetFileService assetsFileService { get; }
+        this.assetsFileService = assetsFileService;
+    }
 
-        public AssetController(IAssetFileService assetsFileService)
-        {
-            this.assetsFileService = assetsFileService;
-        }
+    [SwaggerOperation(Summary = "Endpoint for users to upload file assets.", Description = "Recieves one file in FormData that has the key 'asset'. Returned data value can be accessed at that can be accessed at /assets/{data}")]
+    [HttpPost]
+    public async Task<ActionResult<Response<string>>> PostAsync([FromForm] IFormFile asset)
+    {
+        var newId = await assetsFileService.StoreAsset(asset);
 
-        [SwaggerOperation(Summary ="Endpoint for users to upload file assets.", Description = "Recieves one file in FormData that has the key 'asset'. Returned data value can be accessed at that can be accessed at /assets/{data}")]
-        [HttpPost]
-        public async Task<ActionResult<Response<string>>> PostAsync([FromForm] IFormFile asset)
-        {
-            var newId = await assetsFileService.StoreAsset(asset);
-
-            return Ok(new Response<string> { Data = newId });
-        }
+        return Ok(new Response<string> { Data = newId });
     }
 }
