@@ -1,14 +1,11 @@
-using System;
+namespace Api.Exceptions;
 
-namespace Api.Exceptions
+public class NotFoundException<T> : Exception where T : class
 {
-    public class NotFoundException<T> : Exception where T : class
-    {
-        public NotFoundException() { }
-        public NotFoundException(string message) : base(message) { }
-        public NotFoundException(string message, Exception inner) : base(message, inner) { }
-        protected NotFoundException(
-            System.Runtime.Serialization.SerializationInfo info,
-            System.Runtime.Serialization.StreamingContext context) : base(info, context) { }
-    }
+    public NotFoundException() { }
+    public NotFoundException(string message) : base(message) { }
+    public NotFoundException(string message, Exception inner) : base(message, inner) { }
+    protected NotFoundException(
+        System.Runtime.Serialization.SerializationInfo info,
+        System.Runtime.Serialization.StreamingContext context) : base(info, context) { }
 }
