@@ -1,70 +1,59 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Threading.Tasks;
-using Api.DbModels;
-using Api.DtoModels;
-using Api.Models;
-using Api.Models.Entities;
-using AutoMapper;
+﻿namespace Api.Mappers;
 
-namespace Api.Mappers
+public class AspenMapperProfile : Profile
 {
-    public class AspenMapperProfile : Profile
+    /// <summary>
+    /// Rules to go to and from different object types
+    /// </summary>
+    /// <seealso cref="https://codewithmukesh.com/blog/automapper-in-aspnet-core/"/>
+    public AspenMapperProfile()
     {
-        /// <summary>
-        /// Rules to go to and from different object types
-        /// </summary>
-        /// <seealso cref="https://codewithmukesh.com/blog/automapper-in-aspnet-core/"/>
-        public AspenMapperProfile()
-        {
-            CreateMap<DbEvent, Event>()
-                .ReverseMap();
+        CreateMap<DbEvent, Event>()
+            .ReverseMap();
 
-            CreateMap<Event, DtoEvent>()
-                .ReverseMap();
+        CreateMap<Event, DtoEvent>()
+            .ReverseMap();
 
-            CreateMap<DbPageData, DtoPageData>()
-                .ReverseMap();
+        CreateMap<DbPageData, DtoPageData>()
+            .ReverseMap();
 
-            CreateMap<DbPerson, Person>()
-                .ReverseMap();
+        CreateMap<DbPerson, Person>()
+            .ReverseMap();
 
-            CreateMap<Person, DtoPerson>()
-                .ReverseMap();
+        CreateMap<Person, DtoPerson>()
+            .ReverseMap();
 
-            CreateMap<DbTeam, DtoTeam>()
-                .ReverseMap();
+        CreateMap<DbTeam, DtoTeam>()
+            .ReverseMap();
 
-            CreateMap<DtoTeam, Team>()
-                .ReverseMap();
+        CreateMap<DtoTeam, Team>()
+            .ReverseMap();
 
-            CreateMap<DbTeam, Team>()
-                .ReverseMap();
+        CreateMap<DbTeam, Team>()
+            .ReverseMap();
 
-            CreateMap<DbEvent, Event>()
-                .ReverseMap();
+        CreateMap<DbEvent, Event>()
+            .ReverseMap();
 
-            CreateMap<DtoEvent, Event>()
-                .ReverseMap();
+        CreateMap<DtoEvent, Event>()
+            .ReverseMap();
 
-            CreateMap<DtoRegistration, Registration>()
-                .ReverseMap();
+        CreateMap<DtoRegistration, Registration>()
+            .ReverseMap();
 
-            CreateMap<DbRegistration, Registration>()
-                .ReverseMap();
+        CreateMap<DbRegistration, Registration>()
+            .ReverseMap();
 
-            CreateMap<DbPersonRegistration, PersonRegistration>()
-                .ReverseMap();
+        CreateMap<DbPersonRegistration, PersonRegistration>()
+            .ReverseMap();
 
-            CreateMap<DtoPersonRegistration, PersonRegistration>()
-                .ReverseMap();
+        CreateMap<DtoPersonRegistration, PersonRegistration>()
+            .ReverseMap();
 
-            CreateMap<DbDonation, Donation>()
-                .ReverseMap();
+        CreateMap<DbDonation, Donation>()
+            .ReverseMap();
 
-            CreateMap<DtoDonation, Donation>()
-                .ReverseMap();
-        }
+        CreateMap<DtoDonation, Donation>()
+            .ReverseMap();
     }
 }
