@@ -1,16 +1,12 @@
-﻿using Swashbuckle.AspNetCore.Annotations;
-using System;
+﻿namespace Api.DtoModels;
 
-namespace Api.DtoModels
+public record DtoPersonRegistration
 {
-    public record DtoPersonRegistration
-    {
-        [SwaggerSchema(ReadOnly = true)]
-        public long ID { get; init; }
+    [SwaggerSchema(ReadOnly = true)]
+    public long ID { get; init; }
 
-        public long PersonID { get; init; }
-        public DtoPerson Person { get; init; }
+    public long PersonID { get; init; }
+    public DtoPerson Person { get; init; }
 
-        public DateTime CreatedDate { get; init; }
-    }
+    public DateTime CreatedDate { get; init; }
 }
