@@ -1,23 +1,14 @@
-﻿using Swashbuckle.AspNetCore.Annotations;
-using System;
-using System.Collections.Generic;
-using System.ComponentModel.DataAnnotations;
-using System.ComponentModel.DataAnnotations.Schema;
-using System.Linq;
-using System.Threading.Tasks;
+﻿namespace Api.DtoModels;
 
-namespace Api.DtoModels
+public record DtoTeam
 {
-    public record DtoTeam
-    {
-        [SwaggerSchema(ReadOnly = true)]
-        public long ID { get; init; }
-        [Required]
-        public string Name { get; set; }
-        public string Description { get; init; }
-        public string MainImage {get; init; }
-        public long OwnerID { get; init; }
-        public long EventID { get; init; }
-        public decimal DonationTarget { get; init; }
-    }
+    [SwaggerSchema(ReadOnly = true)]
+    public long ID { get; init; }
+    [Required]
+    public string Name { get; set; }
+    public string Description { get; init; }
+    public string MainImage { get; init; }
+    public long OwnerID { get; init; }
+    public long EventID { get; init; }
+    public decimal DonationTarget { get; init; }
 }
